% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_test3.R
\name{get_items3}
\alias{get_items3}
\title{Item-Ziehung fuer SLT version 3/4}
\usage{
get_items3(
  grammar_type = c("reber_fsg", "loui_fsg"),
  num_items = NULL,
  n_start = 6,
  min_each = 2,
  item_bank = SLT::SLT_item_bank3
)
}
\arguments{
\item{grammar_type}{"reber_fsg" oder "loui_fsg"}

\item{num_items}{Anzahl Items insgesamt. Default NULL: wird automatisch
aus der Anzahl vorhandener Items des jeweiligen grammar_type in
item_bank ermittelt (aktuell: Reber = 36, Loui = 50 - siehe
compile_item_bank3.R). Kann explizit gesetzt werden, z.B. fuer Tests
mit einem kuenstlich verkleinerten Pool.}

\item{n_start}{Anzahl der ersten Items, fuer die Style-Balance erzwungen wird}

\item{min_each}{Mindestanzahl je Style innerhalb der ersten n_start Items}

\item{item_bank}{Quelle der Items, Default SLT::SLT_item_bank3}
}
\description{
Zieht das komplette Item-Pool eines Grammatik-Typs (Style A + Style B,
zu gleichen Teilen) und ordnet die Reihenfolge so an, dass die ersten
n_start Items mindestens min_each Items pro Style enthalten.
}
