% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_test3.R
\name{main_test3}
\alias{main_test3}
\title{Haupt-Timeline fuer SLT version 3/4}
\usage{
main_test3(
  label,
  grammar_type,
  audio_dir,
  dict = SLT::SLT_dict,
  num_items = NULL,
  n_start = 6,
  min_each = 2,
  composer_pairs = list(c(A = "Noa", B = "Sam")),
  autoplay = TRUE,
  ...
)
}
\arguments{
\item{label}{Modul-Label}

\item{grammar_type}{"reber_fsg" oder "loui_fsg"}

\item{audio_dir}{Basis-URL fuer Audio-Dateien}

\item{dict}{psychTestR-Dictionary}

\item{num_items}{Default NULL: wird automatisch aus der Anzahl
vorhandener Items je grammar_type in SLT_item_bank3 ermittelt
(aktuell: Reber = 36, Loui = 50).}

\item{n_start, min_each}{s. get_items3()}

\item{composer_pairs}{Liste mit GENAU EINEM Paar, z.B. list(c(A="Noa", B="Sam"))}

\item{autoplay}{Autoplay fuer Item-Audio}
}
\description{
Haupt-Timeline fuer SLT version 3/4
}
