% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLT.R
\name{SLT}
\alias{SLT}
\title{SLT}
\usage{
SLT(
  num_items = 20L,
  num_blocks = 3L,
  with_welcome = TRUE,
  take_training = FALSE,
  with_finish = TRUE,
  label = "SLT",
  version = 1,
  feedback = SLT_feedback_with_score(dict = SLT::SLT_dict),
  dict = SLT::SLT_dict,
  autoplay = TRUE,
  composer_pairs = NULL,
  ...
)
}
\arguments{
\item{num_items}{(Integer scalar) Number of items in the test.
Fuer version 3/4 wird dieser Wert immer auf 50 gesetzt (25 pro
Grammatik-Typ); ein abweichender Wert erzeugt eine Warnung.}

\item{with_welcome}{(Scalar boolean) Indicates, if a welcome page shall be displayed. Defaults to TRUE}

\item{take_training}{(Logical scalar) Whether to include the training phase. Defaults to FALSE}

\item{with_finish}{(Scalar boolean) Indicates, if a finish (not final!) page shall be displayed. Defaults to TRUE}

\item{label}{(Character scalar) Label to give the SLT results in the output file.}

\item{version}{(Integer scalar) 1 = Markov (Block-Design), 2 = Markov
gamifiziert (Easy/Medium/Hard-Bloecke), 3 = Reber-FSG (deterministisch),
4 = Loui-FSG (deterministisch, Bohlen-Pierce)}

\item{feedback}{(Function) Defines the feedback to give the participant
at the end of the test.}

\item{dict}{The psychTestR dictionary used for internationalisation.}

\item{autoplay}{(Scalar boolean) Indicates whether you want to have autoplay for item pages (instruction pages always not-autoplay)}

\item{composer_pairs}{(List) Composer-Namenspaare pro Block. Bei
version 3/4 wird automatisch EIN Paar (Noa/Sam) verwendet, falls
NULL uebergeben wird (Default).}
}
\description{
This function defines a SLT  module for incorporation into a
psychTestR timeline.
}
