% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{SLT_feedback_with_score}
\alias{SLT_feedback_with_score}
\title{SLT feedback (with score)}
\usage{
SLT_feedback_with_score(dict = SLT::SLT_dict)
}
\arguments{
\item{dict}{The psychTestR dictionary used for internationalisation.}
}
\description{
Here the participant is given textual feedback at the end of the test.
}
\examples{
\dontrun{
SLT_demo(feedback = SLT_feedback_with_score())}
}
