% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLT.R
\name{SLT}
\alias{SLT}
\title{SLT}
\usage{
SLT(
  num_items = 20,
  num_blocks = 3,
  with_welcome = TRUE,
  take_training = FALSE,
  with_finish = TRUE,
  label = "SLT",
  feedback = SLT_feedback_with_score(dict = SLT::SLT_dict),
  dict = SLT::SLT_dict,
  autoplay = TRUE
)
}
\arguments{
\item{num_items}{(Integer scalar) Number of items in the test.}

\item{with_welcome}{(Scalar boolean) Indicates, if a welcome page shall be displayed. Defaults to TRUE}

\item{take_training}{(Logical scalar) Whether to include the training phase. Defaults to FALSE}

\item{with_finish}{(Scalar boolean) Indicates, if a finish (not final!) page shall be displayed. Defaults to TRUE}

\item{label}{(Character scalar) Label to give the SLT results in the output file.}

\item{feedback}{(Function) Defines the feedback to give the participant
at the end of the test.}

\item{dict}{The psychTestR dictionary used for internationalisation.}

\item{autoplay}{(Scalar boolean) Indicates whether you want to have autoplay for item pages (instruction pages always not-autoplay)}
}
\description{
This function defines a SLT  module for incorporation into a
psychTestR timeline.
Use this function if you want to include the SLT in a
battery of other tests, or if you want to add custom psychTestR
pages to your test timeline.
}
\details{
For demoing the SLT, consider using \code{\link{SLT_demo}()}.
For a standalone implementation of the SLT,
consider using \code{\link{SLT_standalone}()}.
}
